+++ rusk-recovery/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- Remove `DEFAULT_SNAPSHOT` const
+
 ## [1.4.0] - 2025-11-06
 
 ### Removed